Placed the file management global methods into a class and exported them as the are fairly useful.
