+++ lib/flutter_sound.dart
@@ -47,3 +47,4 @@ export 'src/ui/recorder_playback_controller.dart'
     show RecorderPlaybackController;
 export 'src/ui/sound_player_ui.dart' show SoundPlayerUI;
 export 'src/ui/sound_recorder_ui.dart' show SoundRecorderUI;
+export 'src/util/file_util.dart' show FileUtil;
